Make operation continuous
