+++ src/components/calculator.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import './calculator.css';
-import { clearValue, getValue } from '../logic/operate';
-import { number, operator } from '../logic/calculate';
+import getValue from '../logic/calculate';
+import { clearValue, number, operator } from '../logic/operate';
 
 /* eslint-disable */
 class Calculator extends React.Component {
@@ -31,13 +31,13 @@ class Calculator extends React.Component {
         <button type="button">7</button>
         <button type="button">8</button>
         <button type="button">9</button>
-        <button type="button" className="operators">*</button>
+        <button type="button" onClick={()=>operator("*")} className="operators">*</button>
         <button type="button">4</button>
         <button type="button">5</button>
         <button type="button">6</button>
         <button type="button" className="operators">-</button>
         <button type="button" onClick={()=>number(1)}>1</button>
-        <button type="button">2</button>
+        <button type="button" onClick={()=>number(2)}>2</button>
         <button type="button">3</button>
         <button type="button" onClick={()=>operator("+")} className="operators">+</button>
         <button type="button" id="zero">0</button>
